requirements.txt
requests
scrapingbee
beautifulsoup4
jmespath
python-dotenv
streamlit
scrapy